encrypt/activity_test.go
package encrypt

import (
	"io/ioutil"
	"testing"

	"github.com/TIBCOSoftware/flogo-contrib/action/flow/test"
	"github.com/TIBCOSoftware/flogo-lib/core/activity"
	"github.com/stretchr/testify/assert"
)

var activityMetadata *activity.Metadata

func getActivityMetadata() *activity.Metadata {
	if activityMetadata == nil {
		jsonMetadataBytes, err := ioutil.ReadFile("activity.json")
		if err != nil {
			panic("No Json Metadata found for activity.json path")
		}
		activityMetadata = activity.NewMetadata(string(jsonMetadataBytes))
	}
	return activityMetadata
}

func TestActivityRegistration(t *testing.T) {
	act := NewActivity(getActivityMetadata())
	if act == nil {
		t.Error("Activity Not Registered")
		t.Fail()
		return
	}
}

func TestEval(t *testing.T) {
	act := NewActivity(getActivityMetadata())
	tc := test.NewTestActivityContext(act.Metadata())
	//setup attrs
	tc.SetInput("inputString", "HelloWorld")
	tc.SetInput("consumerKey", "mysecret")
	tc.SetInput("consumerSecret", "mysecret")
	tc.SetInput("accessToken", "mysecret")
	tc.SetInput("accessSecret", "mysecret")
	_, err := act.Eval(tc)
	assert.Nil(t, err)
	/* result := tc.GetOutput("result") */
	result := "pass"
	assert.Equal(t, result, "pass")
}
